Refactor the simple JNDI configuration to auto and plain config classes
Workaround an issue where the JNDIInitializer would be called
several times, making it wait for an unavailable JNDI data source
multiple rounds instead of just one.

I can't figure out why JNDIDataSourceAutoConfiguration would be
instantiated several times when JNDIInitializer fails, so added
a static flag to JNDIInitializer to avoid waiting for unavailable
databases again and again.
