+++ src/catalog/backends/common/src/main/java/org/geoserver/cloud/autoconfigure/catalog/backend/core/GeoServerBackendAutoConfiguration.java
@@ -6,7 +6,7 @@ package org.geoserver.cloud.autoconfigure.catalog.backend.core;
 
 import org.geoserver.cloud.autoconfigure.geotools.GeoToolsHttpClientAutoConfiguration;
 import org.geoserver.cloud.config.catalog.backend.core.CoreBackendConfiguration;
-import org.geoserver.cloud.config.jndidatasource.JNDIDataSourceAutoConfiguration;
+import org.geoserver.cloud.config.jndi.JNDIDataSourceConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.context.annotation.Import;
@@ -23,6 +23,6 @@ import org.springframework.context.annotation.Import;
  * @see CoreBackendConfiguration
  */
 @AutoConfiguration(
-        after = {GeoToolsHttpClientAutoConfiguration.class, JNDIDataSourceAutoConfiguration.class})
+        after = {GeoToolsHttpClientAutoConfiguration.class, JNDIDataSourceConfiguration.class})
 @Import(CoreBackendConfiguration.class)
 public class GeoServerBackendAutoConfiguration {}
